Adicionando metodo deletar e listar serviço por

+++ src/main/java/com/barbearia/fahcortes/infra/persistence/ServicoRepository.java
@@ -3,5 +3,5 @@ package com.barbearia.fahcortes.infra.persistence;
 import com.barbearia.fahcortes.infra.entities.ServicoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServicoRepository extends JpaRepository<ServicoEntity, Integer> {
+public interface ServicoRepository extends JpaRepository<ServicoEntity, Long> {
 }
